Rename the test helper so that it makes sense inside of a host app
